lerna structure
