+++ packages/udp/package.json
@@ -1,0 +1,50 @@
+{
+  "name": "@kalm/udp",
+  "version": "3.0.0",
+  "description": "UDP transport for Kalm",
+  "main": "./bin/udp.js",
+  "scripts": {
+    "build": "../../scripts/build.sh udp ./src/udp.ts ./bin/udp.js",
+    "test": "mocha ./tests/unit --exit"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/kalm/kalm-js.git"
+  },
+  "keywords": [
+    "multiplex",
+    "network",
+    "optimize",
+    "socket",
+    "websocket",
+    "client",
+    "server",
+    "service",
+    "peer",
+    "micro",
+    "latency",
+    "messaging",
+    "queue",
+    "nagle",
+    "buffer",
+    "protocol",
+    "web",
+    "udp"
+  ],
+  "author": "frederic charette <fredericcharette@gmail.com>",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/kalm/kalm-js/issues"
+  },
+  "homepage": "https://kalm.js.org",
+  "contributors": [
+    "frederic charette <fredericcharette@gmail.com>"
+  ],
+  "browser": {
+    "dgram": false
+  },
+  "devDependencies": {
+    "chai": "~4.2.0",
+    "mocha": "~5.2.0"
+  }
+}
