+++ packages/ws/package.json
@@ -1,0 +1,50 @@
+{
+  "name": "@kalm/ws",
+  "version": "3.0.0",
+  "description": "WebSocket transport for Kalm",
+  "main": "./bin/ws.js",
+  "scripts": {
+    "build": "../../scripts/build.sh ws ./src/ws.ts ./bin/ws.js",
+    "test": "mocha ./tests/unit --exit"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/kalm/kalm-js.git"
+  },
+  "keywords": [
+    "multiplex",
+    "network",
+    "optimize",
+    "socket",
+    "websocket",
+    "client",
+    "server",
+    "service",
+    "peer",
+    "micro",
+    "latency",
+    "messaging",
+    "queue",
+    "nagle",
+    "buffer",
+    "protocol",
+    "web",
+    "ws"
+  ],
+  "author": "frederic charette <fredericcharette@gmail.com>",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/kalm/kalm-js/issues"
+  },
+  "homepage": "https://kalm.js.org",
+  "contributors": [
+    "frederic charette <fredericcharette@gmail.com>"
+  ],
+  "dependencies": {
+    "ws": "^6.1.0"
+  },
+  "devDependencies": {
+    "chai": "~4.2.0",
+    "mocha": "~5.2.0"
+  }
+}
